+++ tests/experiment_test.py
@@ -8,6 +8,8 @@ from beaker import (
     Dataset,
     DatasetNotFound,
     DataSource,
+    ExperimentConflict,
+    ExperimentNotFound,
     ExperimentSpec,
     ImageNotFound,
     ImageSource,
@@ -50,7 +52,11 @@ def test_experiment_spec(client: Beaker, hello_world_experiment_id: str):
 
 
 def test_experiment_create_await_rename(
-    client: Beaker, experiment_name: str, alternate_experiment_name: str, beaker_cluster_name: str
+    client: Beaker,
+    experiment_name: str,
+    alternate_experiment_name: str,
+    beaker_cluster_name: str,
+    hello_world_experiment_name: str,
 ):
     spec = ExperimentSpec(
         tasks=[
@@ -62,13 +68,27 @@ def test_experiment_create_await_rename(
             ),
         ],
     )
+    # Create the experiment.
     experiment = client.experiment.create(experiment_name, spec)
+
+    # Wait for it to finish.
     experiment = client.experiment.await_all(experiment, timeout=60 * 3)[0]
+
+    # Get the logs.
     logs = "".join([line.decode() for line in client.experiment.logs(experiment)])
     assert logs
 
+    # Test experiment conflict error with rename.
+    with pytest.raises(ExperimentConflict):
+        client.experiment.rename(experiment, hello_world_experiment_name)
+
+    # Rename the experiment.
     experiment = client.experiment.rename(experiment, alternate_experiment_name)
     assert experiment.name == alternate_experiment_name
+
+    # Test experiment not found error.
+    with pytest.raises(ExperimentNotFound):
+        client.experiment.get(experiment_name)
 
 
 def test_create_experiment_image_not_found(